pimd,pim6d: use same function to clear GMP groups
Standardize the GMP function for cleaning up all GMP groups in a
interface.
